tests: Cache load encoders when not benchmarking
